Move pkg/fargate to pkg/actions (#3072)
