+++ pkg/actions/cluster/delete.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/weaveworks/eksctl/pkg/actions/fargate"
 	"github.com/weaveworks/eksctl/pkg/cfn/manager"
-	"github.com/weaveworks/eksctl/pkg/fargate"
 
 	api "github.com/weaveworks/eksctl/pkg/apis/eksctl.io/v1alpha5"
 	"github.com/weaveworks/eksctl/pkg/eks"
@@ -24,7 +24,7 @@ func deleteSharedResources(cfg *api.ClusterConfig, ctl *eks.ClusterProvider, cli
 		logger.Debug("failed to check if cluster is operable: %v", err)
 	}
 	if clusterOperable {
-		if err := deleteFargateProfiles(cfg.Metadata, ctl.Provider.WaitTimeout(), ctl); err != nil {
+		if err := deleteFargateProfiles(cfg.Metadata, ctl); err != nil {
 			return err
 		}
 	}
@@ -63,13 +63,12 @@ func handleErrors(errs []error, subject string) error {
 	return fmt.Errorf("failed to delete %s", subject)
 }
 
-func deleteFargateProfiles(clusterMeta *api.ClusterMeta, waitTimeout time.Duration, ctl *eks.ClusterProvider) error {
-	fargateClient := fargate.NewClientWithWaitTimeout(
+func deleteFargateProfiles(clusterMeta *api.ClusterMeta, ctl *eks.ClusterProvider) error {
+	manager := fargate.NewFromProvider(
 		clusterMeta.Name,
-		ctl.Provider.EKS(),
-		waitTimeout,
+		ctl.Provider,
 	)
-	profileNames, err := fargateClient.ListProfiles()
+	profileNames, err := manager.ListProfiles()
 	if err != nil {
 		if fargate.IsUnauthorizedError(err) {
 			logger.Debug("Fargate: unauthorized error: %v", err)
@@ -91,7 +90,7 @@ func deleteFargateProfiles(clusterMeta *api.ClusterMeta, waitTimeout time.Durati
 		// All Fargate profiles must be completely deleted by waiting for the deletion to complete, before deleting
 		// the cluster itself, otherwise it can result in this error:
 		//   Cannot delete because cluster <cluster> currently has Fargate profile <profile> in status DELETING
-		if err := fargateClient.DeleteProfile(*profileName, true); err != nil {
+		if err := manager.DeleteProfile(*profileName, true); err != nil {
 			return err
 		}
 		logger.Info("deleted Fargate profile %q", *profileName)
