+++ pkg/ctl/delete/fargate.go
@@ -6,9 +6,9 @@ import (
 	"github.com/kris-nova/logger"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
+	"github.com/weaveworks/eksctl/pkg/actions/fargate"
 	api "github.com/weaveworks/eksctl/pkg/apis/eksctl.io/v1alpha5"
 	"github.com/weaveworks/eksctl/pkg/ctl/cmdutils"
-	"github.com/weaveworks/eksctl/pkg/fargate"
 )
 
 func deleteFargateProfileWithRunFunc(cmd *cmdutils.Cmd, runFunc func(cmd *cmdutils.Cmd, opts *fargate.Options) error) {
@@ -68,13 +68,13 @@ func doDeleteFargateProfile(cmd *cmdutils.Cmd, opts *fargate.Options) error {
 	}
 
 	clusterName := cmd.ClusterConfig.Metadata.Name
-	awsClient := fargate.NewClientWithWaitTimeout(clusterName, ctl.Provider.EKS(), cmd.ProviderConfig.WaitTimeout)
+	manager := fargate.NewFromProvider(clusterName, ctl.Provider)
 	if cmd.Wait {
 		logger.Info(deletingFargateProfileMsg(clusterName, opts.ProfileName))
 	} else {
 		logger.Debug(deletingFargateProfileMsg(clusterName, opts.ProfileName))
 	}
-	if err := awsClient.DeleteProfile(opts.ProfileName, cmd.Wait); err != nil {
+	if err := manager.DeleteProfile(opts.ProfileName, cmd.Wait); err != nil {
 		return err
 	}
 	logger.Info("deleted Fargate profile %q on EKS cluster %q", opts.ProfileName, clusterName)
